- Add jewel type (object) on noel sale. - Real article limit is on.
- readMe created
- SitePreview added on readMe

minor improvement on UI:
- on search user modal
- add confirmation popUp for suppress sale / article (on validation)
- add confirmation popUp for validate articles transaction
- add confirmation popUp for declare stolen one article
- error color -> red

+++ applicationWebAngular/src/app/article-validation/article-validation.component.ts
@@ -110,10 +110,6 @@ export class ArticleValidationComponent implements OnInit, OnDestroy {
     window.location.reload();
   }
 
-  onFocused(e: any) {
-    // do something
-  }
-
   changeCharacteristics(articleId: number, characteristic: string) {
     this.articles.forEach(article => {
       if (article.id === articleId) { this.articleConcerned = article; }
@@ -138,13 +134,15 @@ export class ArticleValidationComponent implements OnInit, OnDestroy {
   }
 
   onDeleteArticle(articleId: number) {
-    this.articles.forEach(article => {
-      if (article.id === articleId) { this.articleConcerned = article; }
-    });
-    this.articlesService.removeArticle(this.articleConcerned, () => {
-      this.articlesService.getAllArticlesForOneSale(this.saleId, () => {
-        window.location.reload();
+    if (confirm('Voulez-vous vraiment supprimer cet article ?')) {
+      this.articles.forEach(article => {
+        if (article.id === articleId) { this.articleConcerned = article; }
       });
-    });
+      this.articlesService.removeArticle(this.articleConcerned, () => {
+        this.articlesService.getAllArticlesForOneSale(this.saleId, () => {
+          window.location.reload();
+        });
+      });
+    }
   }
 }
